Consolidate region well and titles into a single file for each source
Every well and titles dataset for each state comes from the same source, and sometimes share other fetching or parsing functionality. This moves the two DataSource definitions for each regional source into a single file.
