+++ src/openmethane_prior/sectors/oil_gas/data/nsw_geo.py
@@ -25,8 +25,10 @@ from openmethane_prior.lib.data_manager.parsers import parse_geo
 from openmethane_prior.lib.data_manager.source import ConfiguredDataSource
 
 
+NSW_GEOSCIENCE_WFS_URL="https://public-gs.geoscience.nsw.gov.au/geoserver/wfs"
+
 def fetch_nsw_drillholes(data_source: ConfiguredDataSource):
-    geoserver_wfs = WebFeatureService("https://public-gs.geoscience.nsw.gov.au/geoserver/wfs", version="2.0.0")
+    geoserver_wfs = WebFeatureService(NSW_GEOSCIENCE_WFS_URL, version="2.0.0")
 
     desired_properties = [
         "program",
@@ -77,8 +79,53 @@ def fetch_nsw_drillholes(data_source: ConfiguredDataSource):
 # Source: https://data.nsw.gov.au/data/dataset/coal-seam-gas-borehole
 # Source: https://data.nsw.gov.au/data/dataset/nsw-drillholes-petroleum
 nsw_drillholes_data_source = DataSource(
-    name="nsw-drillholes-csg-petroleum",
+    name="NSW-drillholes-csg-petroleum",
     file_path="NSW-drillholes-csg-petroleum.geojson",
     fetch=fetch_nsw_drillholes,
+    parse=parse_geo,
+)
+
+
+def fetch_nsw_titles(data_source: ConfiguredDataSource):
+    geoserver_wfs = WebFeatureService(NSW_GEOSCIENCE_WFS_URL, version="2.0.0")
+
+    desired_properties = [
+        "tas_id",
+        "title",
+        "holder",
+        "company",
+        "grant_date",
+        "expiry_date",
+        "minerals",
+        "resource",
+        "operation",
+        "geom",
+    ]
+
+    nsw_titles_feature = geoserver_wfs.getfeature(
+        typename="mining-and-exploration:titles_title_granted",
+        srsname="urn:ogc:def:crs:EPSG::4326",
+        propertyname=desired_properties,
+        outputFormat="application/json",
+    )
+
+    features_df = gpd.GeoDataFrame.from_features(json.loads(nsw_titles_feature.read()))
+
+    # only include titles relevant to petroleum production
+    features_df = features_df[features_df["resource"] == "PETROLEUM"]
+    features_df = features_df[features_df["operation"] == "MINING"]
+
+    with open(data_source.asset_path, "w") as asset_file:
+        asset_file.write(features_df.to_json())
+        return data_source.asset_path
+
+
+# Locations of coal seam gas and petroleum production titles in the
+# Australian state of New South Wales, via Geosciences NSW.
+# Source: https://data.nsw.gov.au/data/dataset/nsw-mining-titles
+nsw_titles_data_source = DataSource(
+    name="NSW-titles-csg-petroleum",
+    file_path="NSW-titles-csg-petroleum.geojson",
+    fetch=fetch_nsw_titles,
     parse=parse_geo,
 )
