+++ src/openmethane_prior/sectors/oil_gas/data/qld_gis.py
@@ -17,6 +17,7 @@
 #
 import geopandas as gpd
 import numpy as np
+import pandas as pd
 import restapi # https://github.com/Bolton-and-Menk-GIS/restapi
 
 from openmethane_prior.lib import DataSource
@@ -24,13 +25,13 @@ from openmethane_prior.lib.data_manager.parsers import parse_geo
 from openmethane_prior.lib.data_manager.source import ConfiguredDataSource
 from openmethane_prior.sectors.oil_gas.data.esri_types import map_esri_date_to_str
 
+QLD_SPATIAL_ARCGIS_URL="https://spatial-gis.information.qld.gov.au/arcgis/rest/services"
 
 # Queensland borehole series - REST Service (ArcGIS)
 # https://www.data.qld.gov.au/dataset/queensland-borehole-series/resource/c206a53a-59de-48c2-9544-b7b7323ed5dd
 def fetch_qld_boreholes(data_source: ConfiguredDataSource):
-    # qld_spatial_arcgis = restapi.ArcServer(url="https://spatial-gis.information.qld.gov.au/arcgis/rest/services")
     qld_spatial_boreholes = restapi.MapService(
-        url="https://spatial-gis.information.qld.gov.au/arcgis/rest/services/GeoscientificInformation/Boreholes/MapServer"
+        url=f"{QLD_SPATIAL_ARCGIS_URL}/GeoscientificInformation/Boreholes/MapServer"
     )
 
     boreholes_features = None
@@ -74,8 +75,6 @@ def fetch_qld_boreholes(data_source: ConfiguredDataSource):
 def parse_qld_boreholes(data_source: ConfiguredDataSource):
     boreholes_df = parse_geo(data_source=data_source)
 
-    #
-
     # lease/tenement name, i.e. "PL 100", is more useful than "PL" and 100.0,
     # so combine tenure_no and tenure_type into a single string
     boreholes_df["tenure"] = [
@@ -85,12 +84,69 @@ def parse_qld_boreholes(data_source: ConfiguredDataSource):
 
     return boreholes_df
 
+
 # Locations of all boreholes in the Australian state of Queensland, via the
 # Queensland Open Data Portal.
 # Source: https://www.data.qld.gov.au/dataset/queensland-borehole-series
 qld_boreholes_data_source = DataSource(
-    name="qld-boreholes",
+    name="QLD-boreholes",
     file_path="QLD-boreholes.geojson",
     fetch=fetch_qld_boreholes,
     parse=parse_qld_boreholes,
+)
+
+
+# Petroleum leases - Queensland - REST Service (ArcGIS)
+# https://www.data.qld.gov.au/dataset/queensland-mining-and-exploration-tenure-series/resource/544ecb39-9c25-42f9-9928-b44c7ff05b30
+def fetch_qld_leases(data_source: ConfiguredDataSource):
+    # qld_spatial_arcgis = restapi.ArcServer(url="https://spatial-gis.information.qld.gov.au/arcgis/rest/services")
+    qld_current_leases = restapi.MapService(
+        url=f"{QLD_SPATIAL_ARCGIS_URL}/Economy/MinesPermitsCurrent/MapServer"
+    )
+    qld_historic_leases = restapi.MapService(
+        url=f"{QLD_SPATIAL_ARCGIS_URL}/Economy/MinesPermitsHistoric/MapServer"
+    )
+
+    fields = [
+        "displayname",
+        "permittype",
+        "permitstatus",
+        "approvedate",
+        "expirydate",
+        "permitminerals",
+        "permitpurpose",
+    ]
+
+    leases_layer = qld_current_leases.layer("PL granted")
+    leases_features = leases_layer.query(
+        fields=fields,
+        exceed_limit=True,
+    )
+    historic_layer = qld_historic_leases.layer("Historical petroleum lease")
+    historic_features = historic_layer.query(
+        fields=fields,
+        exceed_limit=True,
+    )
+
+    df = pd.concat([
+        gpd.GeoDataFrame.from_features(leases_features.features),
+        gpd.GeoDataFrame.from_features(historic_features.features),
+    ])
+
+    df["approvedate"] = df["approvedate"].map(map_esri_date_to_str)
+    df["expirydate"] = df["expirydate"].map(map_esri_date_to_str)
+
+    with open(data_source.asset_path, "w") as asset_file:
+        asset_file.write(df.to_json())
+        return data_source.asset_path
+
+
+# Locations of all mining leases in the Australian state of Queensland, via the
+# Queensland Open Data Portal.
+# Source: https://www.data.qld.gov.au/dataset/queensland-mining-and-exploration-tenure-series
+qld_leases_data_source = DataSource(
+    name="QLD-leases",
+    file_path="QLD-leases.geojson",
+    fetch=fetch_qld_leases,
+    parse=parse_geo,
 )
